Add characterHandler.createCharacter
Created an interface called CharacterPrimitive that is the character
model without the foreign references. I used this as the interface
for the properties passed in to the create method. This is different
from how I specified them manually in Story and User. Not sure if
it will come in handy or not, but just something I tried.

I can apply to the other models if I like it or take it out if I
don't.