Replace User model with owner ObjectId property
I want to avoid any server-specific/app-specific code in this
library, so instead of an app-specific User model, owned objects
like Stories and Characters have an owner property that is of
type ObjectId and no specific reference. App can pass and objectId
to whatever document they want to serve as the owner.

Handlers will be responsible for ownership management. Updating
an owned model will require the owner ObjectId and only update if
they match.