Creation handlers return new object without saving
Original handlers were only for testing mongodb.

The flow should be:
* User clicks 'new' and the creation handler returns an owned
object with an _id.
* User saves and an update handler is called that does an usert.
* If User never saves then no harm no foul.

This flow helped immensely when writing front-end because the Redux
store could be indexed by _id and this automatically gives new
objects an _id. Problem is that it adds a request whenever user
hits 'new'.