Add characterHandler.updateCharacter
updateCharacter(_id, props) will check that owner in props owns
the character with the _id and throw an error if not.

Returns updated model object.

Put all non-_id properties in CharacterProps interface so that it
can be specified in function arguments. This means that _id is
supplied via first argument, and all of the other props passed in
the second.